requirements (1).txt
streamlit
python-dotenv
langchain
langchain-community
langchain_huggingface
PyPDF2
